requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
openpyxl==3.1.5
pypdf==5.4.0
Flask-WTF==1.2.2
